Fix CLJ-2619 equivalent bug in claypoole (#60)
There is a memory "leak" for thread futures because the thread bindings are kept around after the future fn runs. We reset the bindings again by the time we run the next future, so it's not a correctness issue. Unfortunately until that next future (which might not happen for a while with large thread pools) we're keeping all the bindings saved for no good reason.